GA:add new feature

+++ cqc-common/src/main/java/cn/wow/common/service/ApplicatService.java
@@ -17,6 +17,6 @@ public interface ApplicatService {
     public List<Applicat> selectAllList(Map<String, Object> map);
     
     // 获取申请人ID
-    public List<Long> selectIds(String applicat_name, String applicat_depart, Integer applicat_org);
+    public List<Long> selectIds(String applicat_name, String applicat_depart, Long applicat_org);
 
 }
