+++ cqc-support/src/main/java/cn/wow/support/web/OtsTaskController.java
@@ -201,7 +201,7 @@ public class OtsTaskController extends AbstractController {
 			String endCreateTime, String task_code, Integer state, Integer draft, Integer atlType, String parts_name,
 			String parts_producer, String parts_producerCode, String startProTime, String endProTime, String matName,
 			String mat_producer, String matNo, String v_code, String v_proAddr, String applicat_name,
-			String applicat_depart, Integer applicat_org, int taskType) {
+			String applicat_depart, Long applicat_org, int taskType) {
 
 		Account account = (Account) request.getSession().getAttribute(Contants.CURRENT_ACCOUNT);
 
@@ -529,7 +529,7 @@ public class OtsTaskController extends AbstractController {
 			String endCreateTime, String task_code, Integer atlType, String parts_name, String parts_producer,
 			String parts_producerCode, String startProTime, String endProTime, String matName, String mat_producer,
 			String matNo, String v_code, String v_proAddr, String applicat_name, String applicat_depart,
-			Integer applicat_org, int taskType) {
+			Long applicat_org, int taskType) {
 
 		// 设置默认记录数
 		String pageSize = request.getParameter("pageSize");
@@ -761,7 +761,7 @@ public class OtsTaskController extends AbstractController {
 			String endCreateTime, String task_code, Integer atlType, String parts_name, String parts_producer,
 			String parts_producerCode, String startProTime, String endProTime, String matName, String mat_producer,
 			String matNo, String v_code, String v_proAddr, String applicat_name, String applicat_depart,
-			Integer applicat_org, int taskType) {
+			Long applicat_org, int taskType) {
 
 		// 设置默认记录数
 		String pageSize = request.getParameter("pageSize");
@@ -963,7 +963,7 @@ public class OtsTaskController extends AbstractController {
 			String endCreateTime, String task_code, Integer atlType, String parts_name, String parts_producer,
 			String parts_producerCode, String startProTime, String endProTime, String matName, String mat_producer,
 			String matNo, String v_code, String v_proAddr, String applicat_name, String applicat_depart,
-			Integer applicat_org, int taskType) {
+			Long applicat_org, int taskType) {
 
 		// 设置默认记录数
 		String pageSize = request.getParameter("pageSize");
